Rewrite stage 7: SDK narrows to types in production
daemonChannelTransport moves to a testonly TestSupport library shared
by both interop test targets; the SDK Server/Client seam conformances
move to the test target's SDKConformances.swift (@retroactive, public
witnesses). Production PreviewsCLI now uses the SDK for data types
only — no Server, no Client, no transports, no import Network. The
daemon_channel_transport_factory lint rule stays (it guards the
test-side channel) with its exclusion repointed; the stale
NetworkTransport rationale in PreviewsMCPApp's dispatchMain comment is
replaced with the surviving one.
