+++ .swiftlint.yml
@@ -59,6 +59,6 @@ custom_rules:
   daemon_channel_transport_factory:
     name: "Daemon-channel transport factory"
     regex: 'NetworkTransport\('
-    excluded: '.*/DaemonChannelTransport\.swift'
+    excluded: '.*/Tests/TestSupport/DaemonChannelTransport\.swift'
     message: "Construct daemon-channel transports via daemonChannelTransport(connection:) — the SDK's default transport heartbeat corrupts framing (see NetworkTransportHeartbeatTests)."
     severity: error
