+++ previewsmcp/PreviewsCLI/MCPClienting.swift
@@ -1,10 +1,10 @@
 import MCP
 
-/// The SDK `Client` surface the CLI actually consumes. Both the SDK client
-/// and `PreviewsMCPClient` conform, so the client parity suite gates the
-/// rewrite differentially and the stage-6 cutover is a construction-site
-/// swap. Tool calls come in via `DaemonToolCalling`, the narrower seam CLI
-/// command bodies already depend on.
+/// The client surface the CLI consumes. `PreviewsMCPClient` conforms in
+/// production; the SDK `Client` conforms test-side (see the test target's
+/// SDKConformances.swift) so the parity suite gates the rewrite
+/// differentially. Tool calls come in via `DaemonToolCalling`, the
+/// narrower seam CLI command bodies already depend on.
 protocol MCPClienting: Actor, DaemonToolCalling {
     @discardableResult
     func onNotification<N: MCP.Notification>(
@@ -14,5 +14,3 @@ protocol MCPClienting: Actor, DaemonToolCalling {
     func connect(transport: any Transport) async throws -> Initialize.Result
     func disconnect() async
 }
-
-extension Client: MCPClienting {}
