+++ previewsmcp/Tests/TestSupport/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@rules_swift//swift:swift.bzl", "swift_library")
+
+swift_library(
+    name = "TestSupport",
+    testonly = True,
+    srcs = glob(["*.swift"]),
+    copts = [
+        "-swift-version",
+        "6",
+    ],
+    module_name = "PreviewsTestSupport",
+    visibility = ["//previewsmcp/Tests:__subpackages__"],
+    deps = [
+        "@swiftpkg_swift_sdk//:MCP",
+    ],
+)
